pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python_turbine_blade_designer"
version = "0.0.1"
authors = [
    { name="Nicola Sorace", email="sorace.n@gmail.com" },
]
description = "Design optimized turbine geometries"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYaml",
    "numpy",
    "pandas",
    "build123d",
    "matplotlib",
]

[project.optional-dependencies]
    pretty_plots = ["seaborn"]
